feat(event-invite): Allow an event invitation to be added to a draft
Summary:
Adds a button in the composer that adds an event card to the draft,
allowing the user to add event details and send the event along with
the draft. When the draft is sent, an event is first created on the
user's calendar from N1. The event_id is then passed to the API's
`/send` endpoint, which sends an ical invite inside the draft to all
recipients.

Adds a new `composer_event_invite` package and modifies the
`SyncbackDraftFilesTask` to also syncback the event (if present)
before sending the draft. Adds a field `event_id` on message JSON
for the `/send` endpoint.

Also fixes `event-header.cjsx` to look at the correct event start
field.

Test Plan: manual

Reviewers: bengotow, evan, juan

Differential Revision: https://phab.nylas.com/D2775